+++ pkg/plugin/aws/iam/user/access/doc.go
@@ -1,4 +1,4 @@
 // Package iam provides a plugin which implements both the rotate.Client and the
 // disable.Client and is used to rotate IAM AWS user accounts and disable
 // inactive access keys associated with those accounts.
-package iam
+package access