+++ pkg/plugin/github/action/secret/doc.go
@@ -1,4 +1,4 @@
 // Package github provides a plugin that implements the rotate.Storage interface
 // to store the results of password rotation into the action secret store of a
 // github project.
-package github
+package secret