+++ pkg/plugin/github/action/secret/github.go
@@ -1,4 +1,4 @@
-package github
+package secret
 
 import (
 	"context"
@@ -10,9 +10,10 @@ import (
 	"time"
 
 	"github.com/google/go-github/v42/github"
+	"golang.org/x/crypto/nacl/secretbox"
+
 	"github.com/zostay/garotate/pkg/config"
 	"github.com/zostay/garotate/pkg/secret"
-	"golang.org/x/crypto/nacl/secretbox"
 )
 
 // secretUpdateAt is the container for last updated date's cache keys.